fmt: use `u32` for `indentation` and `max_width`

+++ compiler/hash-fmt/src/config.rs
@@ -6,9 +6,9 @@
 #[derive(Debug, Clone, Copy)]
 pub struct AstPrintingConfig {
     /// The number of spaces to use when indenting various AST constructs.
-    pub indentation: u8,
+    pub indentation: u32,
 
     /// The maximum allowed width of a line, until the formatting attempts to
     /// wrap a line to the next line.
-    pub max_width: u16,
+    pub max_width: u32,
 }
